Refactor and comment the src directory of the mongodb-service
