+++ mongodb-service/src/idGenerator.ts
@@ -1,3 +1,8 @@
+/**
+ * Helper function that generates valid Mongodb ids. They are not checked against
+ * already existing ones in the dataset so duplicate key errors are theoretically
+ * possible but realistically very unlikely.
+ */
 export function getMongoObjectId() {
     // tslint:disable-next-line:no-bitwise
     const timestamp = (new Date().getTime() / 1000 | 0).toString(16);
