+++ mongodb-service/src/index.ts
@@ -3,8 +3,10 @@
 import logger from './logger/loggerFactory';
 import { start } from './start';
 
-start()
-    .catch((err) => {
-        logger.error(`Error starting server: ${err.message}`);
-        process.exit(-1);
-    });
+/**
+ * The file executed by the "npm run start" command. It starts the API server.
+ */
+start().catch((err) => {
+    logger.error(`Error starting server: ${err.message}`);
+    process.exit(-1);
+});
