+++ mongodb-service/src/logger/loggerFactory.ts
@@ -1,10 +1,11 @@
 import * as log4js from "log4js";
 import { Logger } from "log4js";
 
+/**
+ * Constructs a custom logger.
+ */
 export class LoggerFactory {
-
     public static getLogger(level: string): Logger {
-
         log4js.configure({
             appenders: { 
                 std: { 
@@ -17,7 +18,6 @@ export class LoggerFactory {
             },
             categories: { default: { appenders: ["std"], level: level, enableCallStack: true } }
         });
-
         const logger4 = log4js.getLogger("std");
         return logger4;
     }
