+++ mongodb-service/src/repositories/jobRepository.ts
@@ -5,16 +5,23 @@ import Plot, { IPlotModel } from "../models/plotModel";
 import { AggregationRepository } from "./aggregationRepository";
 import { Repository } from "./repository";
 
+/**
+ * Extends the base Repository class and adds methods unique to Jobs.
+ */
 export class JobRepository extends Repository<IJob> {
     constructor() {
         super(Job);
     }
 
-    // Udate to delete clusters and filters
+    /**
+     * Delete a Job recursively given an id. For Jobs this involves also
+     * deleting recursively the following entities: Aggregations, Plots.
+     * @param id 
+     */
     public async deleteRecursive(id: any): Promise<IJobModel> {
         mongoose.set("useFindAndModify", false);
-        const aggregations: Array<IAggregationModel> = await Aggregation.find({jobId: id}).exec();
-        const plots = await Plot.find({jobId: id}).exec();
+        const aggregations: Array<IAggregationModel> = await Aggregation.find({ jobId: id }).exec();
+        const plots = await Plot.find({ jobId: id }).exec();
 
         const aggregationRepository: AggregationRepository = new AggregationRepository();
 
@@ -32,6 +39,10 @@ export class JobRepository extends Repository<IJob> {
         return await Job.findByIdAndRemove(id).exec();
     }
 
+    /**
+     * Return a colelction of Jobs matching a User id.
+     * @param id 
+     */
     public getjobsByUserId(id: any): Promise<Array<IJobModel>> {
         return Job.find({ userId: id }).exec();
     }
