+++ mongodb-service/src/s3/AWSConfig.ts
