+++ mongodb-service/src/start.ts
@@ -3,18 +3,24 @@
 import { MongoConnector } from './mongodbConnector';
 import { ApiServer } from './api-server';
 
+/**
+ * Instantiate a Mongodb conenction and the API. Handles the starting of the API
+ * Server and the stopping of it and the Mongodb connection.
+ */
 export async function start(): Promise<ApiServer> {
     const mongoConnector = new MongoConnector();
     const apiServer = new ApiServer();
+
     await apiServer.start();
     await mongoConnector.connect();
+
     const graceful = async () => {
         await mongoConnector.disconnect();
         await apiServer.stop();
         process.exit(0);
     };
 
-    // Stop graceful
+    // Stop gracefully.
     process.on('SIGTERM', graceful);
     process.on('SIGINT', graceful);
